requirements.txt
click>=8.0.0
fastapi[standard]
gitingest
google-genai
google-generativeai
httpx>=0.24.0
python-dotenv
pyvis==0.3.2
slowapi
starlette
tiktoken
uvicorn
beautifulsoup4
pymongo